Invent_Manager/requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5.2
distlib==0.3.0
Django==3.0.3
django-crispy-forms==1.9.1
django-el-pagination==3.3.0
django-filter==2.2.0
django-reporting==0.21
filelock==3.0.12
importlib-metadata==1.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
wrapt==1.11.2
zipp==2.2.0
